src/test/java/com/tw/pathashala69/biblioteca/core/auth/UserAuthenticationTest.java
package com.tw.pathashala69.biblioteca.core.auth;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.is;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

class UserAuthenticationTest {

    private User user;

    @BeforeEach
    void setUp() {
        user = mock(User.class);
        UserAuthentication.register(user);
    }

    @Test
    public void shouldRegisterNewUser() {
        assertTrue(UserAuthentication.register(mock(User.class)));
    }

    @Test
    public void shouldLoginUserWithLibraryNumberAndPassword() {
        String libraryNumber = "123-4567";
        String password = "password";
        when(user.authenticate(libraryNumber, password)).thenReturn(true);

        User loggedInUser = UserAuthentication.login(libraryNumber, password);

        assertThat(loggedInUser, is(equalTo(user)));
    }
}